feat(priority_scheduler/priority_queue): rework scheduler to support different scheduling algorithms
The queue configuration now has a type parameter and a priority parameter. The sccheduler now keep
track of the last scheduler item and can be configured to use a different scheduling algorithm

+++ utils/priority_queue/heap/heap.c
@@ -18,7 +18,7 @@ HEAP_TYPE heap_type(heap * hh) {
 	return h->is_min_heap;
 }
 
-int heap_peek(heap * hh) {
+double heap_peek(heap * hh) {
 	_heap * h = hh;
 	return h->array[0].key;
 }
